taskflow-api/requirements.txt
asgiref==3.10.0
attrs==25.4.0
colorama==0.4.6
Django==5.2.7
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
streamlit==1.51.0
tzdata==2025.2
uritemplate==4.2.0
watchdog==6.0.0
wheel==0.45.1
